requirements.txt
flask
openai
langchain
langchain_community
langchain-openai
python-dotenv
